requirements.txt
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pdfkit==1.0.0
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
prettytable~=3.5.0
ciso8601~=2.2.0
requests~=2.28.1
pandas==1.4.4